pallet_native_token_management::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

set_main_chain_scripts

Changes the main chain scripts used for observing native token transfers.

transfer_tokens

Inherent that registers new native token transfer from the Cardano main chain and triggers the handler configured in Config::TokenTransferHandler.