pallet_native_token_management::pallet

Type Alias Initialized

pub type Initialized<T: Config> = StorageValue<_GeneratedPrefixForStorageInitialized<T>, bool, ValueQuery>;

Stores the pallet’s initialization state.

The pallet is considered initialized if its inherent has been successfuly called at least once since genesis or the last invocation of set_main_chain_scripts.

Storage type is [StorageValue] with value type bool.

Aliased Type

struct Initialized<T: Config>(/* private fields */);