pallet_native_token_management

Trait TokenTransferHandler

pub trait TokenTransferHandler {
    // Required method
    fn handle_token_transfer(token_amount: NativeTokenAmount) -> DispatchResult;
}

Interface for user-provided logic to handle native token transfers into the illiquid supply on the main chain.

The handler will be called with the total sum of transfers since the previous partner chain block.

Required Methods

fn handle_token_transfer(token_amount: NativeTokenAmount) -> DispatchResult

New transfer even handler

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.