Crate pallet_native_token_management

Pallet allowing Partner Chains to support movement of their native token from Cardano.

Partner Chains Smart Contracts establish a notion of liquid and illiquid supply of the native token on Cardano, represented as native tokens being either freely available in user accounts or locked under a designated illiquid supply address. Movement of native tokens into the illiquid supply on Cardano signals that an equivalent amount of tokens should be made available on the Partner Chain.

This pallet consumes inherent data containing information on the amount of native tokens newly locked on Cardano and produces an inherent extrinsic to handle their movement. The specific logic releasing the tokens is left for the Partner Chain developer to implement and is configured in the pallet via the TokenTransferHandler.

IMPORTANT: The mechanism implemented via the pallet in this crate is only concerned with the amount of tokens moved and does not attach any metadata to the transfers. In particular it is not a fully-featured token bridge and needs to be combined with a separate sender-receiver metadata channel to implement one.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_native_token_management

Traits

TokenTransferHandler

Interface for user-provided logic to handle native token transfers into the illiquid supply on the main chain.