pallet_native_token_management::pallet

Trait Config

pub trait Config: Config + Config<RuntimeEvent: From<Event<Self>>> {
    type MainChainScriptsOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type TokenTransferHandler: TokenTransferHandler;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type MainChainScriptsOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin for governance calls

type TokenTransferHandler: TokenTransferHandler

Event handler for incoming native token transfers

type WeightInfo: WeightInfo

Weight information for this pallet’s extrinsics

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.